requirements.txt
Django~=4.2
psycopg2~=2.9
python-environ
mysqlclient~=2.1.0
django-debug-toolbar~=4.2.0
djangorestframework~=3.14.0
